+++ Object/PlayerState.cs
@@ -8,30 +8,30 @@
         public string? subCharaID { get; set; }
         public string? mainChaoID { get; set; }
         public string? subChaoID { get; set; }
-        public Int64? numRings { get; set; }
-        public Int64? numBuyRings { get; set; }
-        public Int64? numRedRings { get; set; }
-        public Int64? numBuyRedRings { get; set; }
-        public Int64? energy { get; set; }
-        public Int64? energyBuy { get; set; }
-        public Int64? energyRenewsAt { get; set; }
-        public Int64? mumMessages { get; set; }
-        public Int64? rankingLeague { get; set; }
-        public Int64? quickRankingLeague { get; set; }
-        public Int64? numRouletteTicket { get; set; }
-        public Int64? numChaoRouletteTicket { get; set; }
-        public Int64? chaoEggs { get; set; }
-        public Int64? totalHighScore { get; set; }
-        public Int64? quickTotalHighScore { get; set; }
-        public Int64? totalDistance { get; set; }
-        public Int64? maximumDistance { get; set; }
-        public Int64? dailyMissionId { get; set; }
-        public Int64? dailyMissionEndTime { get; set; }
-        public Int64? dailyChallengeValue { get; set; }
-        public Int64? dailyChallengeComplete { get; set; }
-        public Int64? numDailyChalCont { get; set; }
-        public Int64? numPlaying { get; set; }
-        public Int64? numAnimals { get; set; }
-        public Int64? numRank { get; set; }
+        public ulong? numRings { get; set; }
+        public long? numBuyRings { get; set; }
+        public ulong? numRedRings { get; set; }
+        public long? numBuyRedRings { get; set; }
+        public int? energy { get; set; }
+        public long? energyBuy { get; set; }
+        public long? energyRenewsAt { get; set; }
+        public long? mumMessages { get; set; }
+        public long? rankingLeague { get; set; }
+        public long? quickRankingLeague { get; set; }
+        public long? numRouletteTicket { get; set; }
+        public long? numChaoRouletteTicket { get; set; }
+        public long? chaoEggs { get; set; }
+        public ulong? totalHighScore { get; set; }
+        public ulong? quickTotalHighScore { get; set; }
+        public ulong? totalDistance { get; set; }
+        public ulong? maximumDistance { get; set; }
+        public long? dailyMissionId { get; set; }
+        public long? dailyMissionEndTime { get; set; }
+        public long? dailyChallengeValue { get; set; }
+        public long? dailyChallengeComplete { get; set; }
+        public long? numDailyChalCont { get; set; }
+        public ulong? numPlaying { get; set; }
+        public ulong? numAnimals { get; set; }
+        public short? numRank { get; set; }
     }
 }
