+++ Request/GameRequest.cs
@@ -11,11 +11,11 @@ namespace spikewall.Request
     {
         // Actually means "modifier", misspelling is actually
         // in the game though so we have to keep it like this.
-        public Int64[]? modifire { get; set; }
+        public long[]? modifire { get; set; }
         // This is not always sent, so it needs to be
         // specifically handled during deserialization.
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-        public Int64? tutorial { get; set; }
+        public long? tutorial { get; set; }
     }
 
     /// <summary>
@@ -33,16 +33,16 @@ namespace spikewall.Request
     /// </summary>
     public class QuickPostGameResultsRequest : BaseRequest
     {
-        public Int64? score { get; set; }
-        public Int64? numRings { get; set; }
-        public Int64? numFailureRings { get; set; }
-        public Int64? numRedStarRings { get; set; }
-        public Int64? distance { get; set; }
-        public Int64? dailyChallengeValue { get; set; }
-        public Int64? dailyChallengeComplete { get; set; }
-        public Int64? numAnimals { get; set; }
-        public Int64? maxCombo { get; set; }
-        public Int64? closed { get; set; }
+        public ulong? score { get; set; }
+        public ulong? numRings { get; set; }
+        public ulong? numFailureRings { get; set; }
+        public ulong? numRedStarRings { get; set; }
+        public ulong? distance { get; set; }
+        public long? dailyChallengeValue { get; set; }
+        public long? dailyChallengeComplete { get; set; }
+        public ulong? numAnimals { get; set; }
+        public ulong? maxCombo { get; set; }
+        public long? closed { get; set; }
         public string? cheatResult { get; set; }
     }
 
@@ -52,11 +52,11 @@ namespace spikewall.Request
     /// </summary>
     public class PostGameResultsRequest : QuickPostGameResultsRequest
     {
-        public Int64? bossDestroyed { get; set; }
-        public Int64? chapterClear { get; set; }
-        public Int64? getChaoEgg { get; set; }
-        public Int64? numBossAttack { get; set; }
-        public Int64? reachPoint { get; set; }
+        public long? bossDestroyed { get; set; }
+        public long? chapterClear { get; set; }
+        public long? getChaoEgg { get; set; }
+        public long? numBossAttack { get; set; }
+        public long? reachPoint { get; set; }
 
     }
 
@@ -65,7 +65,7 @@ namespace spikewall.Request
     /// </summary>
     public class MileageRewardRequest : BaseRequest
     {
-        public Int64? episode { get; set; }
-        public Int64? chapter { get; set; }
+        public sbyte? episode { get; set; }
+        public sbyte? chapter { get; set; }
     }
 }
